simplify moderator analysis
make the analysis pipeline more straightforward by computing two separate functions for the continuous moderators and categorical moderators to be fitted in one model respectively (one model for all continuous moderators and one model for all categorical moderators) per trajectory. also added a table that shows a summary of the continuous moderators for all trajectories (table 2 of the paper).